Unbound: object bank, actor/room header, and mesh count lifts
OBJECT_EXCHANGE_BANK_MAX 128 -> 1024 with u16 bank indices (overflow now
logs instead of silently dropping objects); gObjectTable reads are
guarded so object ids beyond the vanilla table are usable. Room header
actor and room counts widen to u16. ActorContext.total was a u8 that
wrapped at 256 so ACTOR_NUMBER_MAX never applied; it is u16 and the cap
is 8192. Mesh display-list counts widen to u32, SHAPE_SORT_MAX is 1024,
and libultraship's texture cache is 8192.

+++ soh/soh/resource/importer/scenecommand/SetMeshFactory.cpp
@@ -18,7 +18,7 @@ std::shared_ptr<Ship::IResource> SetMeshFactory::ReadResource(std::shared_ptr<Sh
     int32_t polyNum = 1;
 
     if (setMesh->meshHeader.base.type != 1) {
-        polyNum = reader->ReadInt8();
+        polyNum = (uint8_t)reader->ReadInt8(); // SOH [Unbound] binary v0 stores a u8 count
         if (setMesh->meshHeader.base.type == 0) {
             setMesh->meshHeader.polygon0.num = polyNum;
         } else if (setMesh->meshHeader.base.type == 2) {
